+++ API.Inventory.CORE/Services/Implementation/RoleInventoryService.cs
@@ -16,16 +16,16 @@ public class RoleInventoryService : IRoleInventoryService
         _unitOfWork = unitOfWork;
         _mapper = mapper;
     }
-    
+
     public async Task<ResponseModel<List<RoleInventoryDto>>> GetAllRoles()
     {
         var response = new ResponseModel<List<RoleInventoryDto>>();
         var roles = await _unitOfWork.RoleInventoryRepository.GetAllRoles();
         
-        response.success = roles.success;
-        response.result = _mapper.Map<List<RoleInventoryDto>>(roles.result);
-        response.errorMessage = roles.errorMessage;
-        
+        response.success = true;
+        response.result = _mapper.Map<List<RoleInventoryDto>>(roles);
+        response.successMessage = "Successfully retrieved roles";
+
         return response;
     }
 
@@ -34,9 +34,9 @@ public class RoleInventoryService : IRoleInventoryService
         var response = new ResponseModel<RoleInventoryDto>();
         var role = await _unitOfWork.RoleInventoryRepository.GetRole(roleId);
         
-        response.success = role.success;
-        response.result = _mapper.Map<RoleInventoryDto>(role.result);
-        response.errorMessage = role.errorMessage;
+        response.result = _mapper.Map<RoleInventoryDto>(role);
+        response.success = true;
+        response.successMessage = "Successfully retrieved role";
         
         return response;
     }
@@ -45,24 +45,56 @@ public class RoleInventoryService : IRoleInventoryService
     {
         var response = new ResponseModel<int>();
         var roleEntity = _mapper.Map<RoleInventory>(role);
+        
+        await _unitOfWork.DbContext.BeginTransactionAsync();
         var result = await _unitOfWork.RoleInventoryRepository.CreateRole(roleEntity);
+        if (result == 0)
+        {
+            await _unitOfWork.DbContext.RollbackTransactionAsync();
+            response.success = false;
+            response.errorMessage = "Error creating role";
+            
+            return response;
+        }
         
-        response.success = result.success;
-        response.result = result.result;
-        response.errorMessage = result.errorMessage;
-        
+        await _unitOfWork.DbContext.CommitTransactionAsync();
+        response.success = true;
+        response.result = result;
+        response.successMessage = "Successfully created role";
+
         return response;
     }
 
     public async Task<ResponseModel<int>> UpdateRole(RoleInventoryDto role)
     {
         var response = new ResponseModel<int>();
-        var roleEntity = _mapper.Map<RoleInventory>(role);
-        var result = await _unitOfWork.RoleInventoryRepository.UpdateRole(roleEntity);
         
-        response.success = result.success;
-        response.result = result.result;
-        response.errorMessage = result.errorMessage;
+        var findRole = await _unitOfWork.RoleInventoryRepository.GetRole(role.RoleInventoryId);
+        if (findRole == null)
+        {
+            response.success = false;
+            response.errorMessage = "Role not found";
+            
+            return response;
+        }
+        
+        var roleProfile = _mapper.Map<RoleInventory>(role);
+        
+        await _unitOfWork.DbContext.BeginTransactionAsync();
+        var result = await _unitOfWork.RoleInventoryRepository.UpdateRole(roleProfile);
+        if (result == 0)
+        {
+            await _unitOfWork.DbContext.RollbackTransactionAsync();
+            response.success = false;
+            response.errorMessage = "Error updating role";
+            
+            return response;
+        }
+        
+        await _unitOfWork.DbContext.CommitTransactionAsync();
+        response.success = true;
+        response.result = result;
+        response.successMessage = "Successfully updated role";
         
         return response;
     }
@@ -70,11 +102,30 @@ public class RoleInventoryService : IRoleInventoryService
     public async Task<ResponseModel<int>> DeleteRole(int roleId)
     {
         var response = new ResponseModel<int>();
-        var result = await _unitOfWork.RoleInventoryRepository.DeleteRole(roleId);
+        var findRole = await _unitOfWork.RoleInventoryRepository.GetRole(roleId);
+        if (findRole == null)
+        {
+            response.success = false;
+            response.errorMessage = "Role not found";
+            
+            return response;
+        }
         
-        response.success = result.success;
-        response.result = result.result;
-        response.errorMessage = result.errorMessage;
+        await _unitOfWork.DbContext.BeginTransactionAsync();
+        var result = await _unitOfWork.RoleInventoryRepository.DeleteRole(roleId);
+        if (result == 0)
+        {
+            await _unitOfWork.DbContext.RollbackTransactionAsync();
+            response.success = false;
+            response.errorMessage = "Error deleting role";
+            
+            return response;
+        }
+        
+        await _unitOfWork.DbContext.CommitTransactionAsync();
+        response.success = true;
+        response.result = result;
+        response.successMessage = "Successfully deleted role";
         
         return response;
     }
